feat: windows build working

+++ src/fec.c
@@ -269,7 +269,7 @@ void writeFloatField(FEC_CONTEXT *ctx, char *filename, const char *extension, in
 // ctx->persistentMemory->line.
 int grabLine(FEC_CONTEXT *ctx)
 {
-  ssize_t bytesRead = readLine(ctx->buffer, ctx->persistentMemory->rawLine, ctx->file);
+  int bytesRead = readLine(ctx->buffer, ctx->persistentMemory->rawLine, ctx->file);
   if (bytesRead <= 0)
   {
     return 0;
